services/database/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
pydantic==2.5.0
python-multipart==0.0.6
python-json-logger==2.0.7
psycopg2-binary==2.9.9
